change amount to parse double

+++ lib/utils/currency_format.dart
@@ -1,5 +1,5 @@
 import 'package:intl/intl.dart';
 
 String currencyDollarFormat(String amount) {
-  return NumberFormat.currency(symbol: '\$').format(int.parse(amount));
+  return NumberFormat.currency(symbol: '\$').format(double.parse(amount));
 }
